backend/requirements.txt
# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.1
websockets>=13.0
python-json-logger>=2.0.7

# Data processing
pandas>=2.2.0
numpy>=2.0.0

# PDF processing
pymupdf4llm>=0.2.8
PyPDF2>=3.0.0

# ScheMatiQ dependencies (for LLM and embedding)
openai>=1.10.0
together>=0.2.0
google-genai>=1.0.0
sentence-transformers>=2.0.0
transformers>=4.40.0
torch>=2.2.0
tiktoken>=0.5.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
PyYAML>=6.0.0

# Cloud Storage
supabase>=2.0.0

# Google Drive/Sheets (optional, for research data collection)
google-api-python-client>=2.100.0
google-auth>=2.25.0
python-dotenv
